[symm_mem] Move all symm mem code into a dedicated folder
[ghstack-poisoned]
