+++ torch/csrc/distributed/c10d/symm_mem/CUDASymmetricMemory.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <ATen/ATen.h>
-#include <torch/csrc/distributed/c10d/CUDASymmetricMemoryTypes.hpp>
+#include <torch/csrc/distributed/c10d/symm_mem/CUDASymmetricMemoryTypes.hpp>
 #include <torch/csrc/distributed/c10d/Store.hpp>
-#include <torch/csrc/distributed/c10d/SymmetricMemory.hpp>
+#include <torch/csrc/distributed/c10d/symm_mem/SymmetricMemory.hpp>
 
 namespace c10d::symmetric_memory {
