+++ torch/csrc/distributed/c10d/symm_mem/CudaDMAConnectivity.cpp
@@ -1,5 +1,5 @@
 #if !defined(USE_ROCM) && defined(PYTORCH_C10_DRIVER_API_SUPPORTED)
-#include <torch/csrc/distributed/c10d/DMAConnectivity.hpp>
+#include <torch/csrc/distributed/c10d/symm_mem/DMAConnectivity.hpp>
 
 #include <c10/cuda/CUDAException.h>
 #include <c10/cuda/driver_api.h>
